chore: extract omnibar target
